* compile.c, cont.c, gc.c, insns.def, iseq.c, iseq.h, process.c,   thread.c, vm.c, vm_core.h, vm_dump.c, vm_eval.c,   vm_insnhelper.c, vm_method.c, template/insns_info.inc.tmpl,   tool/instruction.rb: fixed types.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@25030 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ cont.c
@@ -41,7 +41,7 @@ typedef struct rb_context_struct {
 #endif
     rb_thread_t saved_thread;
     rb_jmpbuf_t jmpbuf;
-    int machine_stack_size;
+    size_t machine_stack_size;
 } rb_context_t;
 
 enum fiber_status {
